Refactor Redux store
- Rename src/store/configureStore.js to src/store/setup.js
- Update path to import statement
